Add record/split_pair and record/disjoint_merge (#1982)
This is preliminary work for the rewriting of native contracts using the
(to be introduced in a coming PR) new immediate/delayed representation.
In particular, in order to efficiently implement the contracts for
static record types, we need a way to split and recombine record which
is efficient and preserves field metadata.

Those primops might be useful to efficiently implement other operations
(record differences, record projection, etc.)