Change custom contract representation (again)
After converting the builtin contracts to the immediate/delayed split
representation, a number of complications appeared. Nothing that can't
be overcome, but requiring unsatisfying ad-hoc treatments or a more
complex contract interface that wasn't easy to explain in the
documentation.

In particular, passing information between the immediate
part and the delayed part is cumbersome. Splitting the logic in two
different functions is arguably harder to understand. Contract
composition, while preserving the immediate and delayed parts, is
cumbersome as well.

In the end, it occurred that switching back to one implementation for
the whole contract, but which can return a value of type `[| 'Ok Dyn,
'Error {..} |]` achieves the same (makes the immediate part and delayed
part separable) while being much simpler both for us maintainers and for
users writing and manipulating contracts. This commit implements this
simplified representation.