Record contracts and apply_as_custom
This commit adapts builtin record contracts, and thus merge in contract
mode as well, to follow the new contract representation: that is, this
commit makes record contract application to return either `'Error` for
immediate errors (the argument isn't a record or there are extra fields)
instead of blaming or an `'Ok`-wrapped result.

We also add a `apply_as_custom` operator, which preserves the
`'Ok`/`'Error` result (as opposed to `apply`) which is useful to
preserve the immediate checks when calling a contract from another
contract.