Add boolean contract combinators to the stdlib
This commit adds the boolean combinators `one_of`, `not` and `all_of` to
the contract module of the stdlib. They are implemented in the obvious
way, by making immediate decisions based on the immediate part of their
contract operands. The limitations and possible work arounds are
detailed in the documentation and in the manual.