Re-introduce unified representation for custom contracts, migrate builtin contracts (#1987)
* Split builtin contracts as immediate/delayed

This commit split the builtin contracts into explicitely separated
immediate and delayed part, as per the contract design introduced in a
previous commit. This incur broad changes, as this impact the internals
module of the stdlib, the stdlib itself, and the code generating
contracts from static types.

* Change custom contract representation (again)

After converting the builtin contracts to the immediate/delayed split
representation, a number of complications appeared. Nothing that can't
be overcome, but requiring unsatisfying ad-hoc treatments or a more
complex contract interface that wasn't easy to explain in the
documentation.

In particular, passing information between the immediate
part and the delayed part is cumbersome. Splitting the logic in two
different functions is arguably harder to understand. Contract
composition, while preserving the immediate and delayed parts, is
cumbersome as well.

In the end, it occurred that switching back to one implementation for
the whole contract, but which can return a value of type `[| 'Ok Dyn,
'Error {..} |]` achieves the same (makes the immediate part and delayed
part separable) while being much simpler both for us maintainers and for
users writing and manipulating contracts. This commit implements this
simplified representation.

* Update stdlib for new ctr repr

Update the stdlib for the new two-in-one custom contract representation
which returns an enum.

* Update tests, examples and doc for new ctr repr

After having introduced a single representation for custom contracts
(instead of a split delayed/immediate), update the various Nickel
snippets in the tests, examples and the documentation to follow this new
API.

* Record contracts and apply_as_custom

This commit adapts builtin record contracts, and thus merge in contract
mode as well, to follow the new contract representation: that is, this
commit makes record contract application to return either `'Error` for
immediate errors (the argument isn't a record or there are extra fields)
instead of blaming or an `'Ok`-wrapped result.

We also add a `apply_as_custom` operator, which preserves the
`'Ok`/`'Error` result (as opposed to `apply`) which is useful to
preserve the immediate checks when calling a contract from another
contract.

* 'Ok -> 'Ok value, use apply_as_custom instead of apply whenever mandated

* Update manual for apply vs apply_as_custom

* Fix usage of renamed of internals

* Fix double contract/custom wrapping failing tests

* Fix some failing tests

* Fix signature of std.contract.custom

* Update test snapshots

* Finish implementation of apply_as_custom

* Small broken test fixes

* Fix typing of contract/custom

* Fix manual snippets

* Fix clippy warning (useless into())

* Apply suggestions from code review

Co-authored-by: jneem <joeneeman@gmail.com>

* Apply suggestions from code review

Co-authored-by: jneem <joeneeman@gmail.com>

* Simplify documentation of contract.apply

---------

Co-authored-by: jneem <joeneeman@gmail.com>