Add new perf measures
This commit use the metrics infrastructure to measure many more
counters, including the number of times each primop is run, each import
is evaluated, the  number of contracts deduped and more. This only
impacts the build with feature `metrics` enabled (disabled by default),
and is proving useful right now to investigate a performance
degradation.