apply_as_custom -> check (#2002)
* apply_as_custom -> check

After bikeshedding, it's been decided in the weekly meeting that we
should use `check` for the new variant of `contract.apply` introduced in
the latest contract system rework.

We couldn't find a meaningful name for this function: `try` and
`try_apply` would be nice, but they are deceptive because they sound
like `apply_as_custom` could catch all contract errors, while it only
makes immediate errors catchable. Also, we might introduce an actualy
`contract.try` at some point, which is able to catch contract errors.

`then` and `and_then` are reflective of how they're used in custom
contract, but that might not be the only possible usage. In the end, we
just needed something different than `apply`, and `check` was the less
bad candidate, with a slight connotation of trying something (as opposed
to `eval`).

* Fix typos in contract.check doc