Fix unsound record contract deduplication (#2042)
* Fix contract equality unsound on records

Contract equality code used to flat out ignore all of the metadata of a
fields, including contract annotations, which is obviously entirely
wrong. This means that `{foo | String}` and `{foo | Number}` was deemed
equal and deduplicated as the same contract.

This commit fixes the issue by comparing the metadata as well. Contract
labels are a subtle matter, and they are the only part that is ignored
by the contract equality code.

* Add comparison of label's type_environment for safety

* Update core/src/typecheck/eq.rs

Co-authored-by: jneem <joeneeman@gmail.com>

* Add integration test for contract dedup unsoundness

---------

Co-authored-by: jneem <joeneeman@gmail.com>