Final import syntax for import with explicit format (#2070)
* Change import syntax and change raw -> text

Change the syntax of import with explicit format specification from the
temporary `import '<Format> <file>` to `import <file> as 'Format`, as
decided with the Nickel team.

Doing so, we also make the various input and export formats name
consistent: we get rid of `raw`, and use `text` instead, which is more
precise. For backward compatibility reason, we still support `raw` as an
alias on the CLI.

The pretty-printer is also fixed: some spaces were previously missing
around import with explicit format specification.

* Apply suggestions from code review

Co-authored-by: jneem <joeneeman@gmail.com>

---------

Co-authored-by: jneem <joeneeman@gmail.com>