Move 'clap' crate dependency behind feature flag in core
As a developer of an app embedding Nickel as a scripting language, I don't want
to be forced to incur the overhead of a third-party commandline-parsing library
(clap) with core Nickel.

The change keeps the new "clap" feature enabled by default in Cargo.toml, to
minimize impact on any users.

See also: https://github.com/tweag/nickel/discussions/2088