Move 'clap' crate dependency behind feature flag in core (#2090)
* Move 'clap' crate dependency behind feature flag in core

As a developer of an app embedding Nickel as a scripting language, I don't want
to be forced to incur the overhead of a third-party commandline-parsing library
(clap) with core Nickel.

The change keeps the new "clap" feature enabled by default in Cargo.toml, to
minimize impact on any users.

See also: https://github.com/tweag/nickel/discussions/2088

* In cli, use core with 'clap' feature enabled

* Don't enable 'clap' feature in core by default

jneem: "I think it's ok to drop clap from the defaults (we'll just have to
remember to bump the crate version). Leaving clap out seems like the right
choice absent backwards-compat constraints, and we don't care too much about
API stability of the core crates."

* Don't wrap ColorChoice in newtype in core

jneem: "maybe it's fine just to export colorchoice::ColorChoice since the crate
is small and it's aligned exactly with what we want."