Remove bytecode-experimental feature
As we move toward a bytecode compiler and a bytecode virtual machine, we
are replacing the left part of the pipeline with the new AST
representation.

The bytecode module was previously gated by an experimental feature,
thea idea being that this feature would enable the whole bytcode
compiler pipeline. However, for now, we only have a new AST
representation, and it's being used in the mainline Nickel parser (and
soon, in the typechecker, etc.). Thus we need access to the new AST
representation by default, and it doesn't make much sense to gate it
behind a feature.

We'll reintroduce the feature once we have a prototype compiler and a
bytecode virtual machine, when it will then make sense to use the
feature to toggle between the legacy tree-walking interpreter and the
new bytecode compiler.