Change function representation to be multi-ary (#2112)
In the new AST, the representation function application has been changed
to be multi-ary by default (as is primop application as well). It just
matches the syntax better and it avoids having to desugar n-ary
application to nested unary-application, saving space and time. The only
cost is that unary application are slighty larger in memory (basically
one word for the size of the slice).

This commit applies the same treatment to function abstraction since the
same arguments apply. Passing by, we also specialize macros like `fun`
and `app` for one argument to avoid vec allocation.