Fix various bug in the new typechecker module
This commit is a backport of many fixes that were done as part of
testing the new typecheck module that operates on the new AST
representation. This commit doesn't yet switch the default typechecker
to the new one, which requires to deeply change the `Cache` structures
and the LSP as well, which are still operating on the old
representation.

It also includes a bunch of new helpers for some AST data structures.