Re-architecturing the cache to accomodate Ast
The new AST borrows from an allocator, which causes a whole lot of
issues in the cache and requires a profound re-architecturing. This
commit start this migration by splitting the cache into smaller, more
manageable parts, and a specific sub-structure that holds the ASTs and
their allocators, which requires a bit of unsafe code.