Add import_paths parameter to Python bindings (#2157)
py-nickel doesn't honor the environment variable `NICKEL_IMPORT_PATH`
nor the CLI argument `--import-path` (obviously, as it doesn't have
access to the latter). This is on purpose, but this means there were no
way to specify additional import paths when running Nickel from Python,
which is is problem.

This commit adds an optional parameter `import_paths`, which is a list
of paths, that can be provided to the `run` function on the Python side.