Get rid of the Terraform/AWS infra for releases
We previously used a bespoke AWS EC2 instance in order to build
statically linked ARM64 linux release artifacts. However, it hasn't
worked in the last 3 releases, and GitHub now provides ARM64 runners out
of the box. It's much simpler to just use them, as we do for X86_64, and
scrap the whole AWS part.