[RFC007] Migration of the typechecker part II - focus on `bytecode::ast::typecheck` (#2170)
* Fix various bug in the new typechecker module

This commit is a backport of many fixes that were done as part of
testing the new typecheck module that operates on the new AST
representation. This commit doesn't yet switch the default typechecker
to the new one, which requires to deeply change the `Cache` structures
and the LSP as well, which are still operating on the old
representation.

It also includes a bunch of new helpers for some AST data structures.

* Update core/src/bytecode/ast/mod.rs

Co-authored-by: jneem <joeneeman@gmail.com>

* Apply suggestions from code review

Co-authored-by: jneem <joeneeman@gmail.com>

* Format merged review suggestion

* Don't allocate a vec: return iterator instead

* Update doc of AstImportResolver

* Remove dead code commented out

* Implement find_path iteratively

---------

Co-authored-by: jneem <joeneeman@gmail.com>