Remove failing assertion
This assertion should hold, but I discovered that it's not the case, for
reasons entirely independent from the current branch: when compiling a
match expression, we assume that we are at eval time, and thus manually
apply the gen_pending_contracts transformation on some compilation
artifact (default value for patterns). However, destructuring is
desugared at transformation time, *before generation of pending
contracts*, so that in that case we pre-transform the default value that
will be transformed again during the normal gen_pending_contracts phase.
Hence the assertion failure.