[Feat] Support annotations on included expressions (#2252)
* Support annotations on include expressions

* Typechecking of include exprs w/ annotations

* Improve a few code comments

* Revert useless diff with previous master

* Fix typechecking of include expr with annotations

* Fix compilation in nls and integration tests

* Remove failing assertion

This assertion should hold, but I discovered that it's not the case, for
reasons entirely independent from the current branch: when compiling a
match expression, we assume that we are at eval time, and thus manually
apply the gen_pending_contracts transformation on some compilation
artifact (default value for patterns). However, destructuring is
desugared at transformation time, *before generation of pending
contracts*, so that in that case we pre-transform the default value that
will be transformed again during the normal gen_pending_contracts phase.
Hence the assertion failure.

* Add first battery of tests

* Formatting

* Fix some free var elaboration errors

* Fix type-fixing in include annotations; remove debug eprintln

* Add missing instantiation when checking annotated includes

* Fix wrong size test in free_var tests

* Fix wrong typecheck error type in test

* Fix missing sep when pretty-printing include exprs

* Fix clippy warning