nix_eval: pass the right base path
For the nix-experimental feature, we didn't provide the correct base
path when instantiating the Nix evaluator, making it impossible to fetch
files from Nix code. This commit fixes the issue, and doing so, adds an
output to the flake to make it easy to build Nickel with the
nix-experimental feature enabled.