into_iter_without_opts -> iter_without_opts
Change `into_iter_without_opts` to not take a record by value. We don't
win much by consuming the record (sparing `NickelValue` and `LocIdent`
clones, which are cheap), and when the record is shared (which is the
most likely), we wastefully duplicate it first before being able to
iterate through it.