refactor: split external state from VM
The virtual machine used to own state that is likely to outlive it. This
makes things awkward, as in the current implementation of e.g. Program,
we need to go through the VM to do a lot of cache operations just
because the VM needs to own it, while those operations have nothing to
do with the VM.

This refactor split this state as a separate context, that the VM
borrows from. The cost is to add a lifetime parameter to VirtualMachine,
but in return, it makes it much simpler to spin local instances of a VM,
instead of threading one VM through many operation mostly just because
we need the cache to continue to exist.

This commit changes a bit the way the stack is unwinded
(`VirtualMachine::reset`). Since we spin a lot more local/temporary
instances, I want to make sure the stack is properly unwinded, which is
easy to forget (by dropping the VM silently), leaving thunks in
blackholed state upon errors. Instead of reseting a VM before the next
eval, this PR incorporates stack unwinding as part of a `Drop`
implementation for the VM, guaranteeing to leave thunks in a consistent
state. The drawback is that we could unwind more than previously (say,
upon a single evaluation that aborts). However, in the happy path
(success), the stack is empty, so the unwinding is virtually free, which
sounds acceptable overall. We can always make a wrapper of
VirtualMachine that doesn't unwind, typically with ManuallyDrop.