refactor: make VM unwind when dropped
Previously, VirtualMachine wouldn't unwind by default when dropped.
Unwinding is probably the best default, as not unwinding can create
undue infinite recursion error if further evaluation happens.

Instead of having VirtualMachine not unwind by default, with a special
wrapper UnwindVirtualMachine for auto-unwinding, we do the converse:
this PR makes VirtualMachine unwind by default, but if for performance
reason one needs to not unwind, a NoUnwindVirtualMachine is provided and
used e.g. in tests.