fix: fix UB (double free) in lenses
This commit adds many missing safety comment, and fix a double-free
error in the lens implementation. We we take a value out of a unique
reference, we need to prevent the original destructor of `NickelValue`
to run. Otherwise, we would return a copy of a value on which a
destructor would have been run. In particular, any heap-allocated data
would be freed/dangling.