fix: fix closurization of recursive records
The migration to `NickelValue` left a broken closurization process for
recursive records. Since the latter can come both fresh out of the
parser, or from prior merging where we must avoid reclosurization, I had
to re-introduce a `closurized` flag (that used to live in `RecordAttrs`
before `NickelValue`) to distinguish both cases.

This is a quick fix, and we really want to refactor `RecRecord` away
from this humongous tuple to a proper anonymous struct. But for now,
that will do.