refactor: introduce FunData & FunPatternData
This commit is a continuation of the compact value representation,
toward the same goal of optimizing the size of the core representations
of the interpreter. After having split WHNFs as compact values, we now
attack the `Term` struct, with the goal of boxing many enum variant data
to bring it back to a reasonable size. This commit starts the work with
functions (`Term::Fun` and `Term::FunPattern`). For now the function
data aren't boxed, as we initially fix the limit to 40 bytes for a
variant. `Term::Fun` fits. We still introduce a separate `FunData`,
which give explicit names to the components, make function data
first-class, and make it easy to box it in the future, should we want to
make `Term` even smaller.