refactor: introduce AppData
While not needed for reducing the size of term, this PR continues the
work of the previous `Term` reduction by introducing a bespoke structure
for an application. This makes `Term` uniform, as `App` was the last
variant constructor with unnamed parameters, and open the door for a
better `TermContent` lens API, where we could decide to look at the
content by reference instead of taking it out by value.