feat: peek term lenses with data()
This commit adds a `data()` helper on term lenses, making it possible to
look in the underlying data by reference without taking ownership,
making lenses more flexible (one can match on a lens but still decide
later on not to take ownership).

Doing so, we also add a `take_unboxed` helper on boxed term variants,
which avoids a `Box` clone - and thus an allocation - for evaluation
patterns where it's common to move out of the content anyway, so that
the box layer is useless.