perf: do not use lenses in eval loop
After some measurement, it turns out it's almost never the case that we
can take advantage of 1-rc values in the main eval loop. Lenses are
still useful for program transformation, but in the main eval loop, it
seems that it's mostly adding a small overhead. This commit goes back to
taking everything by ref.