perf: store thunk data directly in value blocks
This commit fixes a useless indirection when storing a closure/thunk as
a value (or, before that, as a term). We would handle this case by
storing a Thunk, which is itself of the form `Rc<RefCell<ThunkData>>`.
However, value blocks are already reference-counted, so it doesn't make
much sense to store an rc-ed pointer into another rc-ed pointer, for no
obvious benefit.

This commit implements the usual layout for thunks, which is to store
directly the mutable thunk data into the value block, removing an
indirection and an allocation.